requirements.txt
APScheduler==3.10.4
asyncmy==0.2.9
cachebox==4.1.2
dash==2.18.1
DateTime==5.5
fastapi[all]==0.115.0
feffery-antd-charts==0.1.0rc5
feffery-antd-components==0.3.8
feffery-markdown-components==0.2.10
feffery-utils-components==0.2.0rc24
Flask-Compress==1.15
jsonpath-ng==1.5.3
loguru==0.7.2
openpyxl==3.1.5
pandas==2.2.2
passlib[bcrypt]==1.7.4
Pillow==10.4.0
psutil==6.0.0
pydantic-validation-decorator==0.1.2
PyJWT[crypto]==2.8.0
PyMySQL==1.1.1
redis==5.0.7
requests==2.32.3
SQLAlchemy[asyncio]==2.0.31
user-agents==2.2.0
waitress==3.0.0
